cleaned local code plotting af2-spedific data and added af2plots submodule
