+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "alphapulldown/analysis_pipeline/af2plots"]
+	path = alphapulldown/analysis_pipeline/af2plots
+	url = https://gitlab.com/gchojnowski/af2plots.git
